Ensure auto-clearExpireds with different settings
Start auto-clearExpireds when the first view with a defined cache expiry
time is set into the cache and the global "checkExpireds" setting is
used.

For example, when "checkExpireds" is used and a view is set into the
cache, auto-clearExpireds should also start when the global
"cacheExpiry" setting is off and the view's expiry time was set with
the view's `setCacheExpiry` method beforehand.